gateway/requirements.txt
Django==4.2
grpcio>=1.58.0
protobuf==5.29.2
djangorestframework
djangorestframework-simplejwt